refactor: centralize client initialization and improve config
- Move `typesenseClient` initialization to parent component using `useMemo`
- Pass client instance to child components via props
- Remove duplicate client configuration in `EmptyChat` component
